Rename generate-conf.py to scripts/generate_nginx_conf.py + minor refactoring
